Docker image
The process should be able to be run in a Docker container.

This version of the Docker image expects a JSON file with websites to
analyze. It publishes the results to ElasticSearch.

A docker-compose file has been written to create an ElasticSearch
container and a Kibana container. It can be used to publish results
locally with the brienne container.