Better error handling
The process should exit if an error occurs. With async await, a common handler should be used in a "catch"
function. Remember than an async function is run asynchronously and that
it returns a Promise.

A common error handler has been created. It calls process.exit(1) in
case of an error and it prints the stack trace of the error.