Id and date fields
The id field was generated by ElasticSearch. Every time the results were published, it didn't updated the existing results because they couldn't be identified by a unique id.

The id is a hash generated from the id, the base URL and the current URL of the test. Thus, it's unique.

The date field is useful to check when the results have been published to ElasticSearch. it gives an idea of the time that was necessary to perform the data collection.