Introduce the concept of runner
Until now we've been running the data collection with the scripts in a local environment. We're going to use Docker containers to add several workers and even use Kubernetes if needed.

A "local" runner has been created. The main scripts now calls one of the runner configured with the BRIENNE_RUNNER environment variable.