Use a cache
A cache is needed to process the websites faster. Even for the first run, it can cache websites that the parser can visit multiple times.

To evaluate the elements in the page, if we are using the rendered HTML content, we use Cheerio instead of puppeteer. Puppeteer is useful to render the page, but tries to render it again even if we are using the HTML content from the cache.