Use a cursor to distribute tasks
The list of websites was divided among the workers. If one of the workers was processing many invalid websites, the other workers could finish sooner than him and wait a long time, instead of working.

Every worker can now fetch the next website from the cursor. As were are in mono-thread with node.js, the cursor isn't synchronized.