fix: Don't wait before getting the next element
The next element of the cursor was obtained after the creation of a new
page in the browser. In the meanwhile, the other workers were tricked
because the hasNext function told them that there actually was another
element to process.

The cursor.next() function is now called immediately after the
cursor.hasNext() function.